Refactor App.tsx into Components
Break down the App.tsx file into smaller, more manageable components for improved readability and maintainability.
[skip gpt_engineer]
