+++ src/components/LoadingSpinner.tsx
@@ -1,0 +1,7 @@
+const LoadingSpinner = () => (
+  <div className="flex items-center justify-center min-h-screen">
+    <div className="animate-spin rounded-full h-16 w-16 border-t-2 border-b-2 border-forge-orange"></div>
+  </div>
+);
+
+export default LoadingSpinner;
